Separate tracker engines into an internal workspace
