+++ package.json
@@ -21,6 +21,7 @@
     "typescript"
   ],
   "workspaces": [
+    "packages/trackers",
     "packages/core",
     "packages/react-native",
     "packages/web",
@@ -52,9 +53,9 @@
     "build": "npm run build -w supervision-js-core && npm run build -w supervision-js-react-native && npm run build -w supervision",
     "build:js": "npm run build:js -w supervision-js-core && npm run build:js -w supervision-js-react-native && npm run build:js -w supervision",
     "build:types": "npm run build:types -w supervision-js-core && npm run build:types -w supervision-js-react-native && npm run build:types -w supervision",
-    "clean": "npm run clean -w supervision-js-core && npm run clean -w supervision-js-react-native && npm run clean -w supervision",
+    "clean": "npm run clean -w supervision-js-trackers && npm run clean -w supervision-js-core && npm run clean -w supervision-js-react-native && npm run clean -w supervision",
     "dev": "npm run build && concurrently -k -n lib,demo -c cyan,magenta \"npm run dev:lib\" \"npm run dev:demo\"",
-    "dev:lib": "concurrently -k -n core,rn,web -c blue,green,cyan \"npm run dev -w supervision-js-core\" \"npm run dev -w supervision-js-react-native\" \"npm run dev -w supervision\"",
+    "dev:lib": "concurrently -k -n trackers,core,rn,web -c yellow,blue,green,cyan \"npm run dev -w supervision-js-trackers\" \"npm run dev -w supervision-js-core\" \"npm run dev -w supervision-js-react-native\" \"npm run dev -w supervision\"",
     "dev:demo": "npm run dev -w demo",
     "dev:demo-docs": "npm run build && concurrently -k -n demo,docs -c magenta,blue \"npm run dev:demo\" \"npm run docs:dev:server\"",
     "demo:dev": "npm run build && npm run dev:demo",
@@ -93,7 +94,7 @@
     "fixture:sam3:extract": "node tools/sam3-fixture/extract-frames.mjs",
     "fixture:sam3:run": "npm run build -w supervision-js-core && node tools/sam3-fixture/run-sam3.mjs",
     "fixture:geometry:create": "npm run build -w supervision-js-core && node tools/geometry-fixture/create-geometry-fixture.mjs",
-    "typecheck": "npm run typecheck -w supervision-js-core && npm run build -w supervision-js-core && npm run typecheck -w supervision-js-react-native && npm run build -w supervision-js-react-native && npm run typecheck -w supervision && npm run build -w supervision && tsc --noEmit -p demo/tsconfig.json && tsc --noEmit -p examples/vanilla/tsconfig.json && npm run example:react-native:typecheck",
+    "typecheck": "npm run typecheck -w supervision-js-trackers && npm run typecheck -w supervision-js-core && npm run build -w supervision-js-core && npm run typecheck -w supervision-js-react-native && npm run build -w supervision-js-react-native && npm run typecheck -w supervision && npm run build -w supervision && tsc --noEmit -p demo/tsconfig.json && tsc --noEmit -p examples/vanilla/tsconfig.json && npm run example:react-native:typecheck",
     "boundary:check": "node --test tools/package-boundary.test.mjs",
     "test": "vitest run && npm run boundary:check && npm run docs:check && node --test tools/react-runtime-contract.test.mjs tools/release-workflow-contract.test.mjs tools/sam3-fixture/*.test.mjs tools/geometry-fixture/*.test.mjs",
     "package:smoke": "node --test tools/package-smoke.test.mjs",
