+++ packages/core/package.json
@@ -45,11 +45,16 @@
   ],
   "sideEffects": false,
   "scripts": {
-    "build": "npm run clean && npm run build:types && npm run build:js",
-    "build:js": "rollup -c",
-    "build:types": "tsc --emitDeclarationOnly -p tsconfig.json",
+    "build": "npm run build -w supervision-js-trackers && npm run clean && npm run build:types:local && npm run build:js:local",
+    "build:js": "npm run build -w supervision-js-trackers && npm run build:js:local",
+    "build:js:local": "rollup -c",
+    "build:types": "npm run build:types -w supervision-js-trackers && npm run build:types:local",
+    "build:types:local": "tsc --emitDeclarationOnly -p tsconfig.json",
     "clean": "node -e \"fs.rmSync('dist', { recursive: true, force: true })\"",
     "dev": "rollup -c -w",
-    "typecheck": "tsc --noEmit -p tsconfig.json && tsc --noEmit -p tsconfig.test.json"
+    "typecheck": "npm run build:types -w supervision-js-trackers && tsc --noEmit -p tsconfig.json && tsc --noEmit -p tsconfig.test.json"
+  },
+  "devDependencies": {
+    "supervision-js-trackers": "file:../trackers"
   }
 }
